Resolved issues #257, #258, and #259
