+++ src/enums/actions/ActionStatus.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Darling\Rig\enums\actions;
+
+enum ActionStatus
+{
+    case NOT_PROCESSED;
+    case SUCCEEDED;
+    case FAILED;
+}
+
