+++ src/enums/commands/RigCommand.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Darling\Rig\enums\commands;
+
+enum RigCommand: string
+{
+    case DeleteRoute = 'delete-route';
+    case Help = 'help';
+    case ListRoutes = 'list-routes';
+    case NewModule = 'new-module';
+    case NewRoute = 'new-route';
+    case StartServers = 'start-servers';
+    case UpdateRoute = 'update-route';
+    case Version = 'version';
+    case ViewActionLog = 'view-action-log';
+    case ViewReadme = 'view-readme';
+}
